Add uncommon ecdsa oids and handle Crypt::PK::ECC omissions
The OIDS for ecdsa-with-sha1, recommended, and sha2 are missing
but seen in the wild.  Might as well include in the default
table.  (It's always been possible to register them)

Check for the case of Crypt::PK::EC->new dieing in
subjectPublicKeyParams, which can happen if the module loads,
but doesn't understand the key.  Treat it the same way as
if the module isn't present - return undef in the keytype
for the caller to handle.